requirements.txt
Babel==2.3.4
defusedxml==0.4.1
Django==1.9.8
django-allauth==0.26.1
django-bootstrap3==7.0.1
django-datetime-widget==0.9.3
django-disqus==0.5
django-phonenumber-field==1.1.0
gunicorn==19.6.0
mock==2.0.0
oauthlib==1.1.2
pbr==1.10.0
phonenumberslite==7.4.5
python3-openid==3.0.10
pytz==2016.6.1
requests==2.10.0
requests-oauthlib==0.6.2
six==1.10.0
